Configbundit

+++ app/controllers/toys_controller.rb
@@ -2,9 +2,11 @@ class ToysController < ApplicationController
   before_action :set_toy, only: %i[show edit update destroy]
   def index
     @toys = Toy.all
+    @toys = policy_scope(Toy)
   end
 
   def show
+    authorize @toy
     @boxes = Box.all
     @categories = Category.all
     @actions = Action.all
@@ -16,18 +18,21 @@ class ToysController < ApplicationController
     # @action.save
   end
 
-
   def new
     @toy = Toy.new
+    authorize @toy
     @box = Box.find(params[:box_id])
   end
 
   def create
     @box = Box.find(params[:box_id])
     @toy = Toy.new(toy_params)
+    authorize @toy
+    authorize @box
     @toy.box = @box
 
     if @toy.save
+      Action.create!(user: current_user, actionable: @toy, content: "#{current_user.email} à créé le jouet n#{@toy.id}")
       redirect_to toys_path, notice: "Jouet créé avec succès.", status: :see_other
     else
       render :new, status: :unprocessable_entity
@@ -36,10 +41,15 @@ class ToysController < ApplicationController
 
   def edit
     @box = @toy.box
+    authorize @box
+    authorize @toy
   end
 
   def update
+    authorize @toy
     if @toy.update(toy_params)
+      Action.create!(user: current_user, actionable: @toy,
+                     content: "#{current_user.email} à updaté le jouet n#{@toy.id}")
       redirect_to toy_path(@toy), status: :see_other
     else
       render :edit, status: :unprocessable_entity
@@ -47,8 +57,11 @@ class ToysController < ApplicationController
   end
 
   def destroy
+    authorize @toy
     if @toy.destroy
       redirect_to toys_path, notice: "Demande supprimée avec succès."
+      Action.create!(user: current_user, actionable: @toy,
+                     content: "#{current_user.email} à supprimé lejouet #{@toy.id}")
     else
       redirect_to toy_path(@toy), alert: @toy.errors.full_messages.to_sentence
     end
